dashboard/pages/requirements.txt
streamlit
pandas
altair
boto3
python-dotenv
psycopg2
rapidfuzz

